C++ Code: Eliminated all -Wconversion warnings, a few code improvements as a result
Jira: MADLIB-510

Build system:
- Path to Boost is now a system path (to avoid warnings)
- Added "-fdiagnostics-show-option -Wconversion" options to C++ source code

Logistic regression:
- Added protection against oversized independent-variable arrays

Student's t probability function:
- Added checks for negative degrees of freedom

Discrete probability functions:
- Changed arguments to 32-bit integers, so that arguments can be losslessly converted to double, which Boost operates with.

Hypothesis tests:
- one_way_anove(): Changed internal type for groups from uint16_t to uint32_t, in order to match the API

C++ AL:
- utils::MutableReference: deleted operator T&(), which was causing unclear conversion sequences. References cab now be obtained using the ref() method.
- Used size_t and ptrdiff_t types where appropriate

Modules:
- Made several type casts explicit

+++ src/modules/prob/boost.cpp
@@ -124,12 +124,12 @@ DEFINE_CONTINUOUS_PROB_DISTR_2(uniform, double, double)
 DEFINE_CONTINUOUS_PROB_DISTR_2(weibull, double, double)
 
 DEFINE_DISCRETE_PROB_DISTR_1(bernoulli, int32_t, double)
-DEFINE_DISCRETE_PROB_DISTR_2(binomial, int64_t, int64_t, double)
-DEFINE_DISCRETE_PROB_DISTR_1(geometric, int64_t, double)
+DEFINE_DISCRETE_PROB_DISTR_2(binomial, int32_t, int32_t, double)
+DEFINE_DISCRETE_PROB_DISTR_1(geometric, int32_t, double)
 DEFINE_DISCRETE_PROB_DISTR_3(hypergeometric, int32_t, uint32_t, uint32_t,
     uint32_t)
-DEFINE_DISCRETE_PROB_DISTR_2(negative_binomial, int64_t, double, double)
-DEFINE_DISCRETE_PROB_DISTR_1(poisson, int64_t, double)
+DEFINE_DISCRETE_PROB_DISTR_2(negative_binomial, int32_t, double, double)
+DEFINE_DISCRETE_PROB_DISTR_1(poisson, int32_t, double)
 
 } // namespace prob
