+++ src/ports/postgres/dbconnector/SystemInformation_impl.hpp
@@ -22,7 +22,7 @@ inline
 void
 initializeOidHashTable(HTAB*& ioHashTable, MemoryContext inCacheContext,
     size_t inEntrySize, const char* inTabName, uint32_t inMaxNElems) {
-    
+
     if (ioHashTable == NULL) {
         HASHCTL ctl;
         ctl.keysize = sizeof(Oid);
@@ -137,7 +137,7 @@ SystemInformation::typeInformation(Oid inTypeID) {
     // block for performance reasons.
     cachedTypeInfo = static_cast<TypeInformation*>(
         hash_search(types, &inTypeID, HASH_FIND, &found));
-    
+
     if (!found) {
         cachedTypeInfo = static_cast<TypeInformation*>(
             madlib_hash_search(types, &inTypeID, HASH_ENTER, &found));
@@ -155,7 +155,7 @@ SystemInformation::typeInformation(Oid inTypeID) {
             cachedTypeInfo->len = pgType->typlen;
             cachedTypeInfo->byval = pgType->typbyval;
             cachedTypeInfo->type = pgType->typtype;
-            
+
             if (cachedTypeInfo->type == TYPTYPE_COMPOSITE) {
                 // BACKEND: MemoryContextSwitchTo just changes a global
                 // variable
@@ -175,7 +175,7 @@ SystemInformation::typeInformation(Oid inTypeID) {
             madlib_ReleaseSysCache(tup);
         }
     }
-    
+
     return cachedTypeInfo;
 }
 
@@ -183,7 +183,7 @@ SystemInformation::typeInformation(Oid inTypeID) {
  * @brief Get (and cache) information about a PostgreSQL function
  *
  * @param inFuncID The OID of the function of interest
- * @return 
+ * @return
  */
 inline
 FunctionInformation*
@@ -192,21 +192,21 @@ SystemInformation::functionInformation(Oid inFuncID) {
     bool found = true;
     HeapTuple tup;
     Form_pg_proc pgFunc;
-    
+
     // We arrange to look up info about functions only once per series of
     // calls, assuming the function info doesn't change underneath us.
     initializeOidHashTable(functions, cacheContext,
         sizeof(FunctionInformation),
         "C++ AL / FunctionInformation hash table",
         8);
-        
+
     // BACKEND: Since we pass HASH_FIND, this function call will never perform
     // an allocation. There is nothing in the code path that would raise an
     // exception (including oid_hash()), so we are *not* wrapping in a PG_TRY()
     // block for performance reasons.
     cachedFuncInfo = static_cast<FunctionInformation*>(
         hash_search(functions, &inFuncID, HASH_FIND, &found));
-    
+
     if (!found) {
         cachedFuncInfo = static_cast<FunctionInformation*>(
             madlib_hash_search(functions, &inFuncID, HASH_ENTER, &found));
@@ -222,18 +222,19 @@ SystemInformation::functionInformation(Oid inFuncID) {
             cachedFuncInfo->cxx_func = NULL;
             cachedFuncInfo->flinfo.fn_oid = InvalidOid;
             // The number of arguments (excluding OUT params)
-            cachedFuncInfo->nargs = pgFunc->proargtypes.dim1;
+            cachedFuncInfo->nargs
+                = static_cast<uint16_t>(pgFunc->proargtypes.dim1);
             cachedFuncInfo->polymorphic = false;
             cachedFuncInfo->isstrict = pgFunc->proisstrict;
             cachedFuncInfo->secdef = pgFunc->prosecdef;
-            
+
             Oid* allargs;
             // We could use get_func_arg_info() but unfortunately that also
             // copied names and modes
             bool onlyINArguments = false;
             Datum allargtypes = madlib_SysCacheGetAttr(PROCOID, tup,
                 Anum_pg_proc_proallargtypes, &onlyINArguments);
-            
+
             if (onlyINArguments) {
                 allargs = pgFunc->proargtypes.values;
             } else {
@@ -260,12 +261,12 @@ SystemInformation::functionInformation(Oid inFuncID) {
                 // (including OUT params)
                 if (typeInformation(allargs[i])->getType()
                     == TYPTYPE_PSEUDO) {
-                    
+
                     cachedFuncInfo->polymorphic = true;
                     break;
                 }
             }
-            
+
             if (cachedFuncInfo->nargs == 0) {
                 cachedFuncInfo->argtypes = NULL;
             } else {
@@ -276,9 +277,9 @@ SystemInformation::functionInformation(Oid inFuncID) {
                     pgFunc->proargtypes.values,
                     cachedFuncInfo->nargs * sizeof(Oid));
             }
-            
+
             cachedFuncInfo->rettype = pgFunc->prorettype;
-            
+
             // If the return type is RECORDOID, we cannot yet determine the
             // tuple description, even if the function is not polymorphic.
             // For that, the expression parse tree is required.
@@ -289,7 +290,7 @@ SystemInformation::functionInformation(Oid inFuncID) {
             madlib_ReleaseSysCache(tup);
         }
     }
-    
+
     return cachedFuncInfo;
 }
 
@@ -315,15 +316,15 @@ TypeInformation::getTupleDesc(int32_t inTypeMod) {
         // performance).
         TupleDesc pgCachedTupDesc = lookup_rowtype_tupdesc_noerror(oid,
             inTypeMod, /* noerror */ true);
-        
+
         // The tupleDesc is in the cache (RecordCacheArray defined in
-        // typcache.c) even before lookup_rowtype... is called. There is no 
+        // typcache.c) even before lookup_rowtype... is called. There is no
         // need to release the tupleDesc, though we do in order to avoid any
         // side effect.
         ReleaseTupleDesc(pgCachedTupDesc);
         return pgCachedTupDesc;
     }
-    
+
     return NULL;
 }
 
@@ -398,7 +399,7 @@ FunctionInformation::getArgumentType(uint16_t inArgID, FmgrInfo* inFmgrInfo) {
 
         typeID = madlib_get_fn_expr_argtype(inFmgrInfo, inArgID);
     }
-    
+
     return typeID;
 }
 
@@ -423,17 +424,17 @@ FunctionInformation::getReturnType(FunctionCallInfo fcinfo) {
         // sufficient condition for cachedFuncInfo->polymorphic, but not a
         // necessary condition. (A function could have input arguments with
         // pseudo types, but a fixed return type.)
-        
+
         madlib_assert(polymorphic,
             std::logic_error("Logical error: Function returns non-record "
                 "pseudo type but is not polymorphic."));
-        
+
         // This is not a composite type, so no need to pass anything for
         // resultTupleDesc
         madlib_get_call_result_type(fcinfo, &returnType,
             /* resultTupleDesc */ NULL);
     }
-        
+
     return returnType;
 }
 
@@ -451,7 +452,7 @@ FunctionInformation::getReturnTupleDesc(FunctionCallInfo fcinfo) {
         std::runtime_error("Invalid arguments passed to "
             "FunctionInformation::getReturnTupleDesc()."));
 
-    TupleDesc returnTupDesc = tupdesc;    
+    TupleDesc returnTupDesc = tupdesc;
     if (returnTupDesc == NULL) {
         if (rettype == RECORDOID) {
             MADLIB_PG_TRY {
@@ -475,12 +476,12 @@ FunctionInformation::getReturnTupleDesc(FunctionCallInfo fcinfo) {
         } else {
             TypeInformation* cachedTypeInfo
                 = mSysInfo->typeInformation(rettype);
-            
+
             if (cachedTypeInfo->type == TYPTYPE_COMPOSITE)
                 returnTupDesc = cachedTypeInfo->tupdesc;
         }
     }
-    
+
     return returnTupDesc;
 }
 
@@ -509,14 +510,14 @@ FunctionInformation::getFuncMgrInfo() {
         // Check permissions
         if (madlib_pg_proc_aclcheck(oid, GetUserId(), ACL_EXECUTE)
             != ACLCHECK_OK) {
-            
+
             throw std::invalid_argument(std::string("No privilege to run "
                 "function '") + getFullName() + "'.");
         }
 
         // cacheContext will be set as fn_mcxt.
         madlib_fmgr_info_cxt(oid, &flinfo, mSysInfo->cacheContext);
-        
+
         if (!secdef) {
             // If the function is SECURITY DEFINER then fmgr_info_cxt() has
             // set up flinfo so that what we will actually
